+++ core/test/Test.hs
@@ -8,6 +8,7 @@ import Prelude
 import Test.GHCJS qualified as GHCJS
 #endif
 import Test.Canvas qualified as Canvas
+import Test.Cookie qualified as Cookie
 import Test.DriverReentrancy qualified as DriverReentrancy
 import Test.Fork qualified as Fork
 import Test.Hspec (hspec)
@@ -20,6 +21,7 @@ main = hspec $ do
   GHCJS.spec
 #endif
   Canvas.spec
+  Cookie.spec
   DriverReentrancy.spec
   Fork.spec
   NativeDom.spec